requirements_dev.txt
pytest
pytest-django
coverage
codecov
isort==4.3.4
flake8
